feat: move rest button away from the edit button
